pyproject.toml
[project]
name = "captcha-solver"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Volodymyr Klimchuk", email = "ke1wi.ua@gmail.com" }
]
dependencies = [
    "tensorflow>=2.18.0",
    "captcha>=0.6.0",
    "pillow>=11.0.0",
    "matplotlib>=3.9.2",
    "loguru>=0.7.2",
    "tqdm>=4.67.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.7.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/captcha_solver"]
